Fix Keycloak login link and close popup on login
Instead of letting the frontend direct the user to the Keycloak login,
the frontend will call the backend to create the login link. This way,
the backend's Keycloak client can pass the code challenge as a query
parameter to the Keycloak login link.

For more information on how this enhances security against attackers,
research "Keycloak's Proof Key for Code Exchange Code Challenge Method."