Add env variables to Dockerfile for MRN and auth
`NODE_TLS_REJECT_UNAUTHORIZED=0` mutes the error that follows the
initialization of the Keycloak client when the Keycloak admin server is
using a self-signed certificate.

`LD_LIBRARY_PATH` is the default env variable that `node-oracledb' looks
for to access the Oracle client binaries, allow it to run in Thick mode.