Improve SmileNavBar component's typing and responsiveness
This commit hides the user's email and 'Log out' button on mobile-sized
windows. This handles the case where the user's email is too long to fit
on the navbar in the post-login popup window that states 'You are logged
in'. Outside of this scenario, SMILE is a desktop-only application so
this change will not affect the user experience.