Prevent SamplesList from being re-rendered needlessly
The previous implementation of refetchWhereVariables in SamplesPage was
causing double rendering because it was defined inline within the
SamplesList component's props. This meant that this function was
re-created on every render of SamplesPage.

Since SamplesList uses this function in a useEffect hook's dependency
array, each time SamplesPage rendered, it triggered the useEffect in
SamplesList to run again, causing an additional render.